allocatorapi: Add Region param and WithRegion ctx
This patch modifies all the `allocatorapi` public functions to have a
`Region string` required parameter which is then used to create a new
`context.Context`: `api.WithRegion(context.Background(), params.Region)`
and pass it as the API operation context.

Additionally, all the unit tests have been modified to include a request
assertion which ensures that the region path interpolation is used when
the `/platform` APIs are being called.

Last, the `allocator.go` file has been split into multiple files each
containing a single API context and its parameters.
